% Generated by roxygen2 (4.0.0): do not edit by hand
\name{excelTime2char}
\alias{excelTime2char}
\alias{excelTime2num}
\title{Excel Time conversions}
\usage{
excelTime2char(x, debug = FALSE)

excelTime2num(x, debug = FALSE)
}
\arguments{
  \item{x}{a numeric vector}

  \item{debug}{a logical}
}
\value{
a numeric or character vector
}
\description{
Tools to convert to/from Excel time data.
}

