% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letters.R
\name{factorLetters}
\alias{factorLetters}
\title{Turn letter grades into ordered factor}
\usage{
factorLetters(x)
}
\arguments{
\item{x}{a character vector or factor with values that are possible letter grades.}
}
\description{
Turn letter grades into ordered factor
}
\examples{
factorLetters( c("A", "B+", "A-", "B ") )
}

