% Generated by roxygen2 (4.0.1): do not edit by hand
\name{letter2gp}
\alias{letter2gp}
\title{Convert from Letter to Grade Point scale}
\usage{
letter2gp(x)
}
\arguments{
\item{x}{a character vector of grades}
}
\value{
a numeric vector of grade point values
}
\description{
Convert from letter to grade point scale
}
\examples{
letter2gp( c("A", "B-", "HB+", "AU", "IB") )
}

