% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpa.R
\name{letter2gp}
\alias{letter2gp}
\alias{gp2letter}
\title{Convert between Letter and Grade Point scales}
\usage{
letter2gp(x)

gp2letter(x)
}
\arguments{
\item{x}{a character vector of letter grades or a numeric vector of grade points}
}
\value{
a numeric vector of grade point values
}
\description{
Convert between Letter and Grade Point scales
}
\examples{
letter2gp( c("A", "B-", "HB+", "AU", "IB") )
gp2letter( seq(3, 4, by=.1) )

}
