% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{process_scores}
\alias{process_scores}
\title{Collect a series of scores (with drops)}
\usage{
process_scores(data, pattern, as = "hw", drop = 0,
  format = c("percent", "proportion"), ignore.case = TRUE)
}
\arguments{
\item{data}{A data frame containing the gradebook.}

\item{pattern}{A regular expression matched against the names in `data` for selecting
the columns to be aggregated.}

\item{as}{A string naming the new variable to be computed.}

\item{drop}{number of (lowest) items to drop.}

\item{format}{One of `"percent"`, "`proportion"`.}

\item{ignore.case}{A logical indicating whether case should be ignored when matching
`pattern`.}
}
\value{
A data frame
}
\description{
Aggregate scores from matching columns, possibly after dropping some of the lowest scores.
}
