% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{raw2gp}
\alias{raw2gp}
\title{Scale raw grades to gp scores}
\usage{
raw2gp(x, breaks, values = c(4.5, 3.5, 2.5, 1.5, 0.5,
  0)[1:length(breaks)])
}
\arguments{
\item{x}{raw scores to recode}

\item{breaks}{vector of break points between benchmarks}

\item{values}{vector of values to convert break points to.  Interpolation is
used for other values.}
}
\description{
Scale raw grades to gp scores
}
\details{
Both `breaks` and `values` are sorted into decreasing order 
before matching them up.  If the lengths of `breaks` and `values`
differ, the longer is truncated to make them equal length.
}
\examples{
x <- c(50, 63, 78, 91, 54, 17)
raw2gp(x, breaks = seq(50, 100, by = 10))

}
