% Generated by roxygen2 (4.0.1): do not edit by hand
\name{letter2gp}
\alias{gp2letter}
\alias{letter2gp}
\title{Convert between Letter and Grade Point scales}
\usage{
letter2gp(x)

gp2letter(x)
}
\arguments{
\item{x}{a character vector of letter grades or a numeric vector of grade points}
}
\value{
a numeric vector of grade point values
}
\description{
Convert between Letter and Grade Point scales
}
\examples{
letter2gp( c("A", "B-", "HB+", "AU", "IB") )
}

